It kinda works!!

+++ Assets/Core/ChessBot/MovePieces.cs
@@ -461,7 +461,7 @@ namespace ChessEngine
             }
             if (position < 56)
             {
-                byte updatePosition = (byte)(position + 8);
+                int updatePosition = position + 8;
                 while (updatePosition - 8 < 56)
                 {
                     AttackBoard[updatePosition] = true;
@@ -469,7 +469,7 @@ namespace ChessEngine
                     {
                         break;
                     }
-                    moves.Add(new(position, updatePosition));
+                    moves.Add(new(position, (byte)updatePosition));
                     if (HelperFunctions.GetByte(updatePosition, enemyPieces) == 1)
                     {
                         break;
@@ -479,7 +479,7 @@ namespace ChessEngine
             }
             if (position >= 8)
             {
-                byte updatePosition = (byte)(position - 8);
+                int updatePosition = position - 8;
                 while (updatePosition + 8 >= 8)
                 {
                     AttackBoard[updatePosition] = true;
